Extract Masternode Payee data classes
